// src/app/api/model/track.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { YoutubeTrackInfo } from './youtubeTrackInfo';


export interface Track { 
    id?: number;
    title?: string;
    artist?: string;
    youtube?: YoutubeTrackInfo;
    tags?: Array<string>;
}